Move GeometryConfigurationParser & MatrixDecompositionTypeParser
