+++ datadriven/src/sgpp/datadriven/configuration/MatrixDecompositionTypeParser.cpp
@@ -3,9 +3,9 @@
 // use, please see the copyright notice provided with SG++ or at
 // sgpp.sparsegrids.org
 
-#include <sgpp/base/exception/data_exception.hpp>
-#include <sgpp/datadriven/datamining/configuration/MatrixDecompositionTypeParser.hpp>
+#include <sgpp/datadriven/configuration/MatrixDecompositionTypeParser.hpp>
 
+#include <sgpp/base/exception/data_exception.hpp>
 #include <algorithm>
 #include <string>
